% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn.datatree.r
\name{gn.datatree}
\alias{gn.datatree}
\title{gn.datatree}
\usage{
datatree(setupDB)
}
\arguments{
\item{setupDB}{contient une liste des différents paramètres avec les éléments suivants:

 dbname= nom de la base,
 host= hôte,
 port= port,
 name= utilisateur et
 password= mot de passe.}
}
\value{
un plot interactif permettant de visualiser dans la structure de la base.
}
\description{
Cette fonctin permet de visualiser les éléments de la base de donnée accessible pour réaliser une requête.
}
\author{
Rémy Moine <remymoine95@gmail.com>
}
