% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn.setup.r
\name{gn.setup}
\alias{gn.setup}
\title{gn.setup}
\usage{
setupDB<-gn.setup()
}
\arguments{
\item{dbname}{nom de la base de données (type "character")}

\item{host}{nom de l'hôte de votre base de données (type "character")}

\item{port}{port de connexion à votre base de données (type integer)}

\item{user}{nom de l'utilisateur pour se connecter à votre base de données (type "character")}
}
\value{
un object setupDB permettant aux autres fonctions du package d'établir une connexion à votre base de données GeoNature
}
\description{
Cette fonction permet de configurer la connexion à la base de données de votre instance GeoNature. Elle crée un objet nécessaire au bon fonctionnement de nombreuses autres fonctions du package, et doit être jouée au début de votre session de travail.
Si vous supprimez cet objet setupDB, veuillez rejouer la fonction gn.setup() pour reconfigurer la connexion à votre base.
}
\author{
Donovan Maillard <donovan.maillard@gmail.com>
}
