% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_dbextract.R
\name{gn_dbextract}
\alias{gn_dbextract}
\title{gn_dbextract}
\usage{
gn_dbextract(nomdb, id, pw, req)
}
\arguments{
\item{nomdb}{nom de la base deonnée}

\item{id}{identifiant d'accès}

\item{pw}{mot de passe}

\item{req}{la requête à ftraiter sous forme d'un vecteur type "character".}
}
\value{
un data.frame (tableau de données) contenant les résultats de la requête fournit en entrée
}
\description{
Permet d'extraire des données d'une base de données géonature en se basant sur le nom de la base de données seule, les identifiants et un fichier texte fourni.
}
\author{
Remy Moine <remymoine95@gmail.com>
}
